requirements.txt
docopt >= 0.6.1
requirements-parser >= 0.1.0
